s3877256-EEET2574-as1/consumers/requirements.txt
kafka-python
py7zr
nltk
scikit-learn==0.24.1
pandas
cassandra-driver
